infoblox-console/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
python-socketio==5.8.0
python-engineio==4.5.1
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Flask-WTF==1.1.1
pytz==2023.3
colorama==0.4.6
eventlet==0.33.3  # For production WebSocket support